- Do Urgent Task for page: video@show --- Try to look if there’s problem with plug-in: rate-status-plug-in@read/show for rating the shown video.
----- Problem: The notification is being successfully posted. But the problem is that it is being sent to the wrong user (wrong receiver).
The receiver should be the owner of the video and not the user who rated the video.

----- Solution:
The way I solved it is if the notification-msg-id is 6, which is a message about rating a video, then I added an entry to the param-data for creating the parent-notification which is the rateable-item-id.
With that param: rateable-item-id, I can now get the video-obj that has the attribute user-id (the video-uploader-user-id).
With that video-uploader-user-id, I can now set the parent-notification’s notified-user-id.